Fixed one of the leaks.
Fixed some more Lua errors that remained when I converted Python to Lua

Added a 'Shutdown' function for the plugin, so that all of the elements can be garbage collected from Lua before they get destroyed by the other plugins.

Fixed Log.Message

Created the functions for AddEventListener on both Context and Element.

Fixed the LuaEventListener lua function composition to actually compile AND run the string. It works now.