The sample now works completely.
The sample only has one warning upon exit, trying to remove the logo when it has already been destroyed.

Created the DataFormatter class for Lua and the C++ object LuaDataFormatter.

Cleaned up the includes so that any of the exposed API header files only exist in /Include/Rocket/Core/Lua