Hopefully last set of bug fixes/features.
Removed commented out parts for python in the luainvaders sample.

Allowed Colour.rgba to be set as well as read.

Made sure that the functions left the Lua stack as they got it.

Made the DocumentFocus table for flags to be used in Document.Focus function.

Made the Lua function print output to the Log.