Made the proxy tables act just like regular tables
Added some pseudo-metatables named __pairs and __ipairs that types can implement to intercept calls from pairs and ipairs to iterate over values that don't exist in the actual Lua table.

Moved the print override to the GlobalLuaFunctions.cpp file, where the global function pairs and ipairs are overriden to respect the pseudo-metamethods.