add luainvaders to CMakeLists.txt
It compiles in g++ and clang, but make install does not install in a way that makes it able to be run, so I had to run it from the build directory.

There are some funky things going on with msvc and cmake, but there are some pre-built .sln files in the repo for everything, so it shouldn't be a problem until I learn how CMake works.