Added a vc2010 project with instructions
I got tired of converting the old 2008 project each time I've tried to
compile this project. Added a readme file to get a new user up to pace with
the dependency of the freetype library.

There were some mismatch issues with the output filenames after the
conversion, which is all fixed.

The freetype[v]MT.lib is replaced with freetype[v].lib. MT is for
multi-threaded non-dll projects, and the latter is for multi-threaded dll
projects. This is the better choice since LibRocket is a dll project. Mixing
runtimes is bad mojo which can lead to subtle and hard to find bugs. Better
safe than sorry.

Also took the chance of upgrading the freetype library to v2.4.10. The
readme states which options to change in order to alter the desired version,
so its a quick fix for new and old users.

The new project is in its own directory in the build folder. All paths have
been updated for this and should work as expected. The support folder is
still in the paths and is also updated.

All final output files end up in the rep_root/bin folder. Debug versions
appended with '_d'.