Add natural display dimensions to decorators based on spritesheet source scaling and dp-ratio.
This means that when applying a dp-ratio on a context, tiles (eg. corners of a ninepatch or tiled box decorator) will scale accordingly. Tiles will also respect the 'resolution' property in a spritesheet which defines the scale the image is meant to be displayed at, effectively the reverse of the effect of dp-ratio.

The old hack of scaling the edges of the ninepatch decorator by 1.0 to make it respond to the dp-ratio no longer has any effect, this is now effectively the default.