Fix an issue where some elements could end up rendered at the wrong offset after scrolling, see #230.
Modify Element::DirtyOffset so that all descendents are dirtied. It is possible that a child is dirty while a grandchild is not, eg. if the grandchild is hovered over thereby calling Element::GetAbsoluteOffset to resolve the offset *and* the grandchild's offset_parent is its grandparent. Then previously the grandchild would never be dirtied even when its grandparent was dirtied, which again would result in the wrong offset during Element::GetAbsoluteOffset.