Refactor shell, introduce backends
A backend is a combination of a renderer and a platform. The shell is now only used for common functions specific to the included samples, most of the old shell has been replaced by the new backends. The backend used for the samples can be selected during CMake configuration, allowing all the samples to run on any backend.

All samples and tests have been updated to work with the backends interface, which is a very light abstraction over all the different backends. The old SFML and SDL samples are removed, instead, they can be selected as backends that work on all the samples.

A new GLFW backend has been added, along with the SFML and SDL platforms ported from the old samples. The old macOS shell has been removed as it used a legacy API that is no longer working on modern Apple devices. Now the samples build again on macOS using one of the windowing libraries such as GLFW or SDL.

See the Backends section in 'readme.md' for more details.

Development commits:
- Start reworking shell.
- Refactoring shell, replaces old shell
- Update Windows include
- Update visual tests, unit tests, and benchmarks for new shell interface
- Update demo sample for new shell interface
- Minor include fix for Win32-GL2 backend
- Fix win32 platform backend
- Add backend for X11
- Fix tests
- Rename include headers for windows and xlib
- Update Win32/GL2 backend
- Rename shell function names
- PlatformExtensions default file extension
- Update basic samples and tutorials for new shell
- Update invader sample for the new shell
- Update LuaInvader sample for the new shell
- Add SDL2 backend
- Add SDL backend with native SDL renderer
- Add SFML backend with OpenGL renderer. Update SDL backend and GL2 renderer.
- Refactor shell.
- Remove old shell code.
- Remove SDL2, SDL2renderer, and SFML samples. These are replaced by backends that work with all the samples.
- Move Backends to project root directory.
- Clean up CMake code for the shell and backends.
- Fix SVG and Lottie samples.
- Fix include, update CI
- Additional include fixes, update CI
- Minor fix
- Add macOS library to CMake
- Minor CMake fixes
- Add GLFW backend
- Some cleanup to shell and backends
- Remove GL2 stencil command
- Add backends readme and update samples readme
- Refactor backends, remove global state, move the main loop into the samples, call directly into the backend interface instead of through the shell.