Add type converters for all the big integer types.
Solves a problem with emscripten where fetching size_t from Variant always returned zero. In Emscripten, size_t is long instead of int (both 32-bit), and no converter was specified for this type. Then again long collides with int64_t on other platforms, so instead of just adding long we replace the sized types with converters for all the big integer types.

This issue affected the data binding sample, where clicking on the invaders did not register a hit.