Refactor ComputedValues to reduce memory usage.
Reduces the byte size of ComputedValues to less than half of previous.

Split the computed values into common values, rare values, and inherited values. Tried a few different optimization strategies, such as using copy-on-write for the inherited and rare values. Didn't seem to make a noticable difference so for now I dropped this to avoid the added complexity. Might be worthwhile to look into this some more later.

Retrieving computed values are now done through function calls instead of member objects, so it touches a lot of files for that reason.