Fix Lua object garbage collection bug (#340)
As discovered by Ishq, multiple extant userdata Lua objects sometimes
had the same underlying 'obj' pointer (pointer to the underlying object
which is used as the key in the "DO NOT TRASH" table). Which led to
crashes. Lua does not immediately garbage-collect dead objects. This
means that for the non-owning objects (gc=false), another object could
be potentially be created at the same address before garbage collection
occurs, thereby clobbering the DO NOT TRASH entry.

The fix: use the address of the userdata, which is guaranteed to be
unique, as the DO NOT TRASH key.