Add Node class
Element now derives from a new Node class. Node is responsible for managing the DOM tree, while Element is more closely tied to the RML nodes.

This opens up the possibility of having different nodes than just Element.

This change alone should be functionally equivalent and backward compatible. The intention is to later align naming and functionality to the HTML DOM. This will involve some breaking changes later on. Functions that are known to change in functionality later have been marked deprecated.