Only allow a single render interface, remove the ability to set per-context render interfaces [breaking change]
Now, only a single render interface can be set, and it must be set globally before the call to Rml::Initialise. Furthermore, the render interface cannot be changed while the library is initialised, and must live until after the call to Rml::Shutdown. If these conditions are not met, the library will submit error messages to the user in most cases.

This change is motivated by several reasons:

- The feature added considerable complexity, since several types of render state needed to store their state per render interface.
- This complexity would have become more pronounced with upcoming changes for advanced render effects.
- It was easy to make bugs where render resources (e.g. textures) would be freed in the wrong render interface.
- It was difficult to test for the presence of the above bugs.
- No tests existed already, it is very likely that bugs were already present.