SVG decorators, caching, and content cropping
The `<svg>` element has gotten a new attribute `crop-to-content`. If present, the displayed SVG viewbox will be cropped to the content of the SVG, i.e. scaling up the content until it reaches the edges of the texture.

The SVG decorator can be instantiated using the syntax:
```
decorator: svg("path.svg" <crop>?) <paint-area>?;
```
The `crop` property can optionally be set to the `crop-to-content` keyword, which has the same effect as the equivalent `<svg>` attribute. A second example document is added to the SVG sample, demonstrating the decorator.

Finally, there is now a cache for SVG documents. Each SVG document will be stored in the cache as long as it is in use, otherwise the document will be released. For each SVG document, we also cache textures from it based on their resolution and color. Textures will be reused if all properties that determine their generated bitmap match.