Add Input Method Editor (IME) support (#541)
Adds `TextInputHandler` abstract class which users can implement to integrate IME support. Users can override the abstract methods and install it with `Rml::SetTextInputHandler` or pass it during context construction `Rml::CreateContext`.

The text input handler is notified when a text field is activated, and receives a pointer to a `TextInputContext` interface. The library implements this class for its text input elements. Through this class, the user can query the text field for information such as selection and bounding box, and apply the composition text directly directly into the text field.

IME support is currently implemented in the Win32 backend which uses the system IME in Windows.

New sample `rmlui_sample_ime` added for IME testing. Loads system fonts to support a wide selection of languages and color emojis.

---------

Co-authored-by: Michael Ragazzon <michael.ragazzon@gmail.com>