Make the client area always tied to the padding area, textarea elements now clip to their padding area
This removes the support for customizing the client area of an element, and fixes it to its padding area. This makes it easier to understand what the client area is, and is also in line with how CSS defines this term. Previously, the client area was also used to define how the element was clipped. Instead, this is now made more explicit with a separate "clip area" that can be customized for each element.

Normally, clipping occurs on the padding area of the element. The clip area is introduced mainly for single-line text inputs, as we still want to clip to the content area here.

On the other hand, textarea elements previously also clipped to the content area, but this is undesirable and not according to CSS specifications. Now, the textarea elements clip to padding, just like normal elements.

The 'invaders.rcss' textarea decorators previously assumed clipping to the content area. This would now lead to content being placed on top of the decorator when scrolling, which looks strange. Instead, span the decorator across the border-area of the element, and make a transparent border as a placeholder. This better describes the nature of the decorator, since after all it attempts to display a border. Since clipping now acts on the padding area, the content will be clipped when it is scrolled into the border area, where the decorator now displays the borders. Furthermore, adjust the position of the textarea scrollbars accordingly, and improve their looks generally.