ElementHandle: Keep the target anchored after being positioned and sized, resolves #637
This commit changes the how the position and size of handle targets are determined. If an element has all of its inset (top/right/bottom/left) properties set, this determines the size. Previously, we converted this to a width/height and always used top/left when sizing and positioning the target element. This is no longer the case, we now keep the target anchored to the same sides as it was anchored to before positioning and sizing.

Now, when first sizing and moving the target, and then resizing its container, the element can now still size itself to match the new dimensions.

This commit additionally fixes a couple issues, where the element jolts some distance at drag start:

1. When the target's offset parent had a border.
2. When set to relative positioning and placed away from the top-left corner.