Handle element: Introduce `edge_margin` attribute to constrain position to containing block edges, see #631
This attribute can take any length or percentage, which specifies the minimum distance between the target and the edges of its containing block. These constraints will be satisfied while dragging the handle, for both position and size targets.

By default, this value is set to `0px`, which means that handle targets will now be constrained exactly to the edges of their containing block. The value can also be specified by up to four space-separated values in box order (top/right/bottom/left), determining the minimum distance to each side. Negative values are allowed, which enables movement outside the edges of the containing block. Percentages are resolved against the size of the target element. The value "none" can be used to remove any constraints.