Allow observer pointers after Rml::Shutdown, but be more assertive about elements
It is rather innocent for users to keep objects derived from Rml::EventListener around after shut down, and it can be a hassle to work around that. So instead, allow that situation and rather make observer pointers garbage collect themselves if they are destroyed after Rml::Shutdown.

On the other hand, keeping an Rml::Element around after Rml::Shutdown will most likely lead to a crash. Be more assertive when detecting this and make it an error.