Text widget: Remove copy of scroll offset to avoid scroll jump after layout
The scroll offset member was not always kept in sync with the element's actual scroll offset. This sometimes caused trouble during re-layout when an old value of scroll offset was applied, causing the text widget contents to jump during layout.

The member seems to serve no good purpose anymore, so we simply remove it. Possibly, the idea was that setting the scroll offset after layout helped restore the scroll offset in case it was temporarily clamped in the middle of formatting. However, this should no longer be happening now that there is more control over scroll clamping behavior.

This change may also have improved IME positioning.