Skip rounding in flexbox layout, allow fractional sizes of flex items
Previously the flex item offset and size were snapped to the pixel grid. This made it so that flex items could have a slightly different size than their inner content. This could further cause a single pixel gap between the child of a flex item and that child's immediate cousin.