Fix loading font faces through Lua (#764)
When the font face index was first proposed (#720), it was added as a second
parameter. It was later refactored to be the last parameter. However, when this
change was made, it was not reflected in the Lua plugin.

I took this opportunity to introduce an additional parameter for the Lua
function: the fallback face flag. Keeping it as the second parameter does not
break anything, as that is what the face index argument actually was for Lua
before.