Allow layout caching, even under max-content constraint
This introduces a new rule that max-content constraint should also format things the same way, as long as the parameters passed to the cache is the same too.

If we want to stray from this rule, we could later add separate caches for max-content and the final content. Or alternatively, say that we can use the final cache for max-content, but never the max-content cache for final layout.

This change further improves nested flexbox layout (from 18 to 15).