Fix structural data views (data-for) not working with data-model on the body element, see #790
The issue being that the data model is not available on the body element while its XML is being parsed, until it is attached to the context's root element. The normal data views are instantiated only after attaching the document in this case. However, this did not work for structural data views, they could only be attached during XML parsing.

This commit addresses the issue by allowing structural data views also to be attached while attaching the document to the context root.

An alternative fix could be to make the data model available during XML parsing, as the context is known at that point. Both approaches could supplement each other.

It seems the previously reported issue with data-model not working with body templates was the same underlying issue as addressed here. A test has been added, and this issue is considered fixed as well.